// frontend/src/openapi/models/Session.ts
export interface Session {
    /**
     *
     * @type {string}
     * @memberof Session
     */
    readonly id: string;
    /**
     *
     * @type {string}
     * @memberof Session
     */
    readonly externalId: string;
    /**
     *
     * @type {string}
     * @memberof Session
     */
    readonly name: string;
    /**
     *
     * @type {string}
     * @memberof Session
     */
    readonly project: string;
    /**
     *
     * @type {string}
     * @memberof Session
     */
    readonly capability: string | null;
    /**
     *
     * @type {string}
     * @memberof Session
     */
    readonly capabilityName: string | null;
    /**
     *
     * @type {number}
     * @memberof Session
     */
    readonly traceCount: number;
    /**
     *
     * @type {number}
     * @memberof Session
     */
    readonly spanCount: number;
    /**
     *
     * @type {number}
     * @memberof Session
     */
    readonly firstSpanNs: number | null;
    /**
     *
     * @type {number}
     * @memberof Session
     */
    readonly lastSpanNs: number | null;
    /**
     *
     * @type {number}
     * @memberof Session
     */
    readonly sessionScore: number | null;
    /**
     *
     * @type {number}
     * @memberof Session
     */
    readonly totalTokens: number | null;
    /**
     *
     * @type {number}
     * @memberof Session
     */
    readonly totalCost: number | null;
    /**
     *
     * @type {string}
     * @memberof Session
     */
    readonly model: string | null;
    /**
     *
     * @type {Date}
     * @memberof Session
     */
    readonly createdAt: Date;
}